+++ src/REvoDesign/shortcuts/shortcuts.py
@@ -455,7 +455,8 @@ def dump_sidechains(
     dpi: int = 150,
     ray: bool = True,
     hide_mesh: bool = True,
-    neiborhood: int = 3,
+    neighborhood: int = 3,
+    reorient: bool = True,
     recenter: bool = False,
 ):
     """
@@ -469,7 +470,9 @@ def dump_sidechains(
       dpi (int, optional): DPI of the image. Defaults to 150.
       ray (bool, optional): If True, uses ray tracing. Defaults to True.
       hide_mesh (bool, optional): If True, hides the mesh. Defaults to True.
-      neiborhood (int, optional): Zoom with neiborhood. Defaults to 3.
+      neighborhood (int, optional): Zoom with neighborhood. Defaults to 3.
+      reorient (bool, optional): If True, re-orients the residue. Defaults to True to prevent automatic orientation.
+        Useful when user wants to dump the residue they just focused on.
       recenter (bool, optional): If True, re-centers the resiude. Defaults to False.
     Returns:
       None
@@ -487,6 +490,7 @@ def dump_sidechains(
 
     # disable all groups
     cmd.disable(' or '.join(all_groups))
+    cmd.refresh()
 
     for sel in sele:
         # ensure the current group selection is enabled
@@ -499,10 +503,11 @@ def dump_sidechains(
         cmd.disable(' or '.join(all_models))
 
         # orient to get pose in the right orientation
-        if neiborhood and neiborhood > 0:
-            cmd.orient(f'{sel} or byres {sel} around {neiborhood}')
+        if reorient and neighborhood and neighborhood > 0:
+            cmd.orient(f'{sel} or byres {sel} around {neighborhood}')
 
         for m in all_models:
+            cmd.refresh()
             print(f"Dumping PNG for {m} ...")
             cmd.enable(m)
             cmd.show("sticks", m)